Fix #7: Import succeeds w/o content.xml
We have to cater to missing or corrupt content.xml files, by not
importing at all.

+++ classes/ImportService.php
@@ -55,6 +55,9 @@ class ImportService
     public function import()
     {
         $root = simplexml_load_file($this->xmlFilename);
+        if (!$root) {
+            return false;
+        }
         $this->purgeContents();
         foreach ($root->page as $page) {
             $this->createPage($page);
